+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Releases
 
-## Version 3.0.101 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
+## Version 3.0.102 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
+- Enhanced Senses "Sight Group" now adds to PERCEPTION skill roll. [#1288](https://github.com/dmdorman/hero6e-foundryvtt/issues/1230)
+
+
+## Version 3.0.101
 - We no longer test/verify with FoundryVTT V11. We encourage you upgrade to FoundryVTT v12.
 - Talent/Skill/Perks as powers now toggle. [#1288](https://github.com/dmdorman/hero6e-foundryvtt/issues/1230)
 - Improved/fixed Penalty Skill Levels. You can now have more than one PSL for different attacks.